pyproject.toml
[tool.poetry]
name = "home-qr-generator"
version = "0.1.0"
description = ""
authors = ["Kotet <kotet0418@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
qrcode = "^7.3.1"
Pillow = "^9.0.1"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
